maybe-working firehose refactor
